renovate weeks 2-3-4

+++ 04-operator-overloading/2-brackets-operator/units/IntList.cpp
@@ -2,18 +2,18 @@
 #include <stdexcept>
 
 void IntList::operator=(int value) {
-    for (uint i=0; i<this->numInts; ++i)
+    for (size_t i=0; i<this->numInts; ++i)
         theInts[i] = value;
 }
 
-const int IntList::operator[](uint index) const {
+const int IntList::operator[](size_t index) const {
     cout << "c ";
     if (index >= numInts)
         throw out_of_range("Array index out of bounds: "+to_string(index));
     return theInts[index];
 }
 
-int& IntList::operator[](uint index) {
+int& IntList::operator[](size_t index) {
     cout << "r ";
     if (index >= numInts)
         throw out_of_range("Array index out of bounds: "+to_string(index));
